// Core/src/wallet/core/model/codes/statusCodes.ts
export enum StatusCodes
{
    OK = 200,
    Created = 201,
    NoContent = 204,
    BadRequest = 400,
    Unauthorized = 401,
    Forbidden = 403,
    NotFound = 404,
    RequestTimeout = 408,
    Conflict = 409,
    LengthRequired = 411,
    InternalServerError = 500,
    NotImplemented = 501,
    ServiceUnavailable = 503,
    HTTPVersionNotSupported = 505,
}